refactor(codebase): consolidate git credential value-objects
Cleanup from the /simplify pass over the credential-leak fix.

- Move GitAuthEnv to codebase/clients/base.py beside its sibling
  GitEgressCredential (both are Basic-oauth2 credential value-objects
  produced by RepoClient methods), and drop base.py's deferred
  self-import of it. Relocate its unit tests to test_base.py.
- Extract a shared _basic_oauth2() helper used by both for_token methods,
  so the "both transport paths present the same credential shape" security
  invariant is held by one function instead of two hand-written copies.
- Simplify the _git_local env overlay: as_env() already carries the
  prompt-disabling vars, so only the no-credential branch sets them
  (removes the redundant double-set and the misleading "wins" comment).
- Collapse the repeated GitAuthEnv.for_token(...).as_env() expressions in
  the GitLab client tests behind a _expected_clone_env helper.

No behavior change; full suite green.
