+++ daiv/codebase/clients/base.py
@@ -9,6 +9,7 @@ from dataclasses import dataclass
 from enum import StrEnum
 from functools import cached_property
 from typing import TYPE_CHECKING, Any
+from urllib.parse import urlparse
 
 from pydantic import SecretStr
 
@@ -31,8 +32,6 @@ if TYPE_CHECKING:
     from github import Github
     from gitlab import Gitlab
 
-    from codebase.clients.utils import GitAuthEnv
-
 logger = logging.getLogger("daiv.clients")
 
 
@@ -47,16 +46,27 @@ class WebhookSetupResult(StrEnum):
     SKIPPED = "skipped"
 
 
+def _basic_oauth2(token: str) -> str:
+    """Build the ``Basic base64("oauth2:<token>")`` credential DAIV presents for git-over-HTTPS.
+
+    Single source of truth for both transport paths — the egress-proxy-injected header
+    (:class:`GitEgressCredential`) and the local-mode clone/subprocess env header
+    (:class:`GitAuthEnv`) — so the remote always sees the identical credential shape (the security
+    invariant those two carry) rather than two hand-written copies that could drift."""
+    encoded = base64.b64encode(f"oauth2:{token}".encode()).decode()
+    return f"Basic {encoded}"
+
+
 @dataclass(frozen=True)
 class GitEgressCredential:
     """Egress-proxy contribution for a repo's git platform: which host to allow and the
     ``Authorization`` header to inject so git-over-HTTPS in the sandbox is authenticated.
 
     The injected header is the sandbox's *only* git credential — the seeded clone carries none
-    (see :class:`codebase.clients.utils.GitAuthEnv`). ``value`` is ``None`` when no token was
-    provisioned — either the platform needs none (e.g. SWE eval's public repos) or one was required
-    but could not be minted. The host is still returned so reachability works; auth, if required,
-    then fails at the git layer."""
+    (see :class:`GitAuthEnv`). ``value`` is ``None`` when no token was provisioned — either the
+    platform needs none (e.g. SWE eval's public repos) or one was required but could not be minted.
+    The host is still returned so reachability works; auth, if required, then fails at the git
+    layer."""
 
     host: str
     header: str = "Authorization"
@@ -64,13 +74,63 @@ class GitEgressCredential:
 
     @classmethod
     def for_token(cls, *, host: str, token: str | None) -> GitEgressCredential:
-        """Build a credential injecting ``Authorization: Basic base64("oauth2:<token>")`` —
-        the same shape DAIV's clone auth env uses. ``value`` is ``None`` when ``token`` is falsy."""
-        value = None
-        if token:
-            encoded = base64.b64encode(f"oauth2:{token}".encode()).decode()
-            value = SecretStr(f"Basic {encoded}")
-        return cls(host=host, value=value)
+        """Build a credential injecting ``Authorization: Basic base64("oauth2:<token>")``.
+        ``value`` is ``None`` when ``token`` is falsy."""
+        return cls(host=host, value=SecretStr(_basic_oauth2(token)) if token else None)
+
+
+@dataclass(frozen=True)
+class GitAuthEnv:
+    """A credential + prompt-disabling environment overlay for git-over-HTTPS, carried without the
+    token ever touching argv (``ps``-visible) or ``.git/config`` (which is seeded into the sandbox).
+
+    The credential is held in :class:`~pydantic.SecretStr` — like the sibling
+    :class:`GitEgressCredential` — so it never appears verbatim in a ``repr``, a log line, or a
+    Sentry stack-local; the plaintext is materialised only by :meth:`as_env`, which callers invoke at
+    the innermost subprocess/clone boundary. Build via :meth:`for_token`.
+    """
+
+    config_key: str
+    """The ``http.<origin>.extraheader`` config key. Keeps the clone URL's scheme and port because
+    git matches ``http.<url>.*`` by prefix: a mismatch would silently send no credential."""
+
+    header: SecretStr
+    """The ``Authorization: Basic base64("oauth2:<token>")`` header value — the same shape the egress
+    proxy injects (see :meth:`GitEgressCredential.for_token`)."""
+
+    @classmethod
+    def for_token(cls, clone_url: str, token: str) -> GitAuthEnv:
+        """Build the overlay for ``clone_url``'s origin authenticated with ``token``.
+
+        Args:
+            clone_url: The repository's credential-less HTTP(S) clone URL.
+            token: The token to authenticate with.
+        """
+        parsed = urlparse(clone_url)
+        return cls(
+            config_key=f"http.{parsed.scheme}://{parsed.netloc}/.extraheader",
+            header=SecretStr(f"Authorization: {_basic_oauth2(token)}"),
+        )
+
+    def as_env(self) -> dict[str, str]:
+        """Materialise the overlay as git subprocess environment variables (plaintext credential).
+
+        ``GIT_CONFIG_{COUNT,KEY_0,VALUE_0}`` apply the command-scoped ``extraheader``.
+        ``GIT_TERMINAL_PROMPT=0`` **and** ``GIT_ASKPASS=""`` together disable every prompt path so a
+        *rejected* credential fails fast with ``could not read Username`` — one of
+        :func:`core.utils.is_git_auth_error_text`'s markers, so the clone-retry self-heal and
+        push-failure classifier keep recognising auth errors. Both are needed: with only
+        ``GIT_TERMINAL_PROMPT=0`` git still falls back to an inherited ``SSH_ASKPASS`` GUI helper and
+        hangs; the empty ``GIT_ASKPASS`` is non-null (short-circuiting that fallback chain) yet empty
+        (so nothing is executed), leaving only the disabled terminal prompt.
+        """
+        return {
+            "GIT_CONFIG_COUNT": "1",
+            "GIT_CONFIG_KEY_0": self.config_key,
+            "GIT_CONFIG_VALUE_0": self.header.get_secret_value(),
+            "GIT_TERMINAL_PROMPT": "0",
+            "GIT_ASKPASS": "",
+        }
 
 
 class RepoClient(abc.ABC):
@@ -150,8 +210,6 @@ class RepoClient(abc.ABC):
         The shared shape lives here — derive the host from the clone URL, then build a
         ``Basic oauth2:<token>`` credential. Platforms supply only the token via
         :meth:`_git_egress_token` (overriding this method is unnecessary)."""
-        from urllib.parse import urlparse
-
         host = urlparse(repository.clone_url).hostname
         if not host:
             return None
@@ -167,13 +225,11 @@ class RepoClient(abc.ABC):
         (push/fetch/ls-remote), or ``None`` for platforms whose remotes need no credential
         (e.g. SWE eval's public repos).
 
-        The clone persists no credential (see :class:`codebase.clients.utils.GitAuthEnv`), so
-        sandbox-disabled runs must overlay it on each git subprocess instead. Resolved at call
-        time — the token is short-lived, so it is minted/fetched per publish rather than pinned at
-        clone time. A ``None`` return also covers the case where a credential *was* required but
-        could not be minted; the debug log distinguishes that from the no-credential-needed case."""
-        from codebase.clients.utils import GitAuthEnv
-
+        The clone persists no credential (see :class:`GitAuthEnv`), so sandbox-disabled runs must
+        overlay it on each git subprocess instead. Resolved at call time — the token is short-lived,
+        so it is minted/fetched per publish rather than pinned at clone time. A ``None`` return also
+        covers the case where a credential *was* required but could not be minted; the debug log
+        distinguishes that from the no-credential-needed case."""
         token = self._git_egress_token(repository)
         if not token:
             logger.debug(
